requirements.txt
agate==1.13.0
agate-dbf==0.2.3
agate-excel==0.4.1
agate-sql==0.7.2
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
attrs==23.2.0
audioread==3.0.1
Authlib==1.6.0
av==15.1.0
axios==0.4.0
babel==2.17.0
beautifulsoup4==4.12.3
blinker==1.9.0
Brotli==1.1.0
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.2.1
cmake==3.28.1
colorama==0.4.6
contourpy==1.3.2
cryptography==42.0.5
csvkit==2.1.0
cycler==0.12.1
Cython==3.0.8
dbfread==2.0.7
decorator==4.4.2
Deprecated==1.2.14
docker==7.1.0
einops==0.7.0
et-xmlfile==1.1.0
fastapi==0.115.12
ffmpy==0.6.0
filelock==3.13.4
Flask==3.1.2
fonttools==4.58.2
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google_search_results==2.4.2
gradio==4.44.1
gradio_client==1.3.0
gradiologin==0.1.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.22.2
idna==3.7
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_resources==6.5.2
intel-openmp==2021.4.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
kornia==0.7.2
kornia_rs==0.1.3
# Editable install with no version control (ktem==0.10.7)
-e c:\users\hi\documents\01_brandschutzgpt\kotaemon\libs\ktem
langchain==0.3.25
langchain-core==0.3.65
langchain-text-splitters==0.3.8
langsmith==0.3.45
lazy_loader==0.4
leather==0.4.0
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==6.0.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matrix-client==0.4.0
mdurl==0.1.2
mkl==2021.4.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
mutagen==1.47.0
networkx==3.3
numba==0.60.0
numpy==1.26.3
olefile==0.47
omegaconf==2.3.0
opencv-python==4.10.0.84
openpyxl==3.1.5
orjson==3.10.18
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.2
parsedatetime==2.6
pefile==2024.8.26
pillow==10.3.0
pinecone==7.0.2
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
pluggy==1.6.0
pooch==1.8.2
proglog==0.1.10
psutil==5.9.8
pyasn1==0.6.1
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
PyGithub==2.3.0
Pygments==2.19.1
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyJWT==2.8.0
PyNaCl==1.5.0
pyparsing==3.2.3
PySimpleGUI==5.0.6
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.12
python-slugify==8.0.4
pytimeparse==1.1.8
pytrends==4.9.2
pytz==2024.1
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.1
regex==2024.4.16
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9
ruff==0.11.13
safetensors==0.4.3
scikit-learn==1.5.2
scipy==1.13.0
selenium==4.24.0
semantic-version==2.10.0
seo-keyword-research-tool==0.1.9
setuptools==69.0.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
sympy==1.13.1
tail==0.1.12
tbb==2021.12.0
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.19.1
tomlkit==0.12.0
torchsde==0.2.6
tqdm==4.66.2
trampoline==0.1.2
transformers==4.40.1
trio==0.26.2
trio-websocket==0.11.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2024.1
tzlocal==5.3.1
urllib3==2.4.0
uv==0.9.14
uvicorn==0.34.3
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.4
wget==3.2
wheel==0.42.0
wrapt==1.16.0
wsproto==1.2.0
xlrd==2.0.2
yarl==1.9.4
youtube-dl==2021.12.17
yt-dlp==2025.11.12
zstandard==0.23.0
